# swagger.yml
openapi: 3.0.0
servers:
  - url: /api/v1
info:
  description: 'This will be a living spec, if you find errors or inconsistencies please raise these concerns. But **please do not change the spec without consulting with the team**'
  version: 0.0.8
  title: SENG302 Team 500 Travel EA API Spec
...
paths:
...
  '/travellers/passports':
    get:
      tags:
        - travellers
      summary: Gets a list of all valid passports
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully retrieved passports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/passports'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '403':
          description: >-
            Profile has not been completed (Only if viewing other users
            passports)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  '/travellers/nationalities':
    get:
      tags:
        - travellers
      summary: Gets a list of all valid nationalities
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully retrieved nationalities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/nationalities'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '403':
          description: >-
            Profile has not been completed (Only if viewing other users
            nationalities)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  schemas:
...
    error:
      type: object
      properties:
        error:
          type: string
          description: An informative error message (to help with debugging)
...
    passport:
      type: object
      properties:
        passportId:
          type: integer
        passportCountry:
          type: string
    passports:
      type: array
      items:
        $ref: '#/components/schemas/passport'
    nationality:
      type: object
      properties:
        nationalityId:
          type: integer
        nationalityName:
          type: string
    nationalities:
      type: array
      items:
        $ref: '#/components/schemas/nationality'
...
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer